Update codebase to upstream revision a626d865c0b27d583b92ce359ad484d86aabdce1.

+++ samples/net/dhcpv4_client/src/main.c
@@ -38,7 +38,7 @@ static void start_dhcpv4_client(struct net_if *iface, void *user_data)
 }
 
 static void handler(struct net_mgmt_event_callback *cb,
-		    uint32_t mgmt_event,
+		    uint64_t mgmt_event,
 		    struct net_if *iface)
 {
 	int i = 0;
